check balance route added

+++ models/user.go
@@ -10,12 +10,12 @@ import (
 )
 
 type Users struct {
-	ID        int64     `gorm:"column:id; primary_key; auto_increment" json:"id"`
+	ID        int       `gorm:"column:id; primary_key; auto_increment" json:"id"`
 	Uuid      uuid.UUID `gorm:"column:uuid; unique_index; default: null" json:"uuid"`
 	Name      string    `gorm:"column:name;" json:"name"`
 	Phone     Phone     `gorm:"column:phone; unique_index" json:"phone"`
 	PIN       string    `gorm:"column:pin; default: null" json:"-"`
-	AccountId int64     `gorm:"column:account_id" json:"account_id"`
+	AccountId int       `gorm:"column:account_id" json:"account_id"`
 	Active    bool      `gorm:"column:active; default:true" json:"active"`
 	BirthDate time.Time `gorm:"column:birth_date; default: null" json:"birth_date"`
 	CreatedAt time.Time `gorm:"default: CURRENT_TIMESTAMP" json:"-"`
